refactor(silo): rename subdirectory roaring to roaring_util
